time-series/requirements.txt
joblib==0.13.2
numpy==1.16.3
pandas==0.24.2
pkg-resources==0.0.0
py-dateutil==2.2
python-dateutil==2.8.0
pytz==2019.1
scikit-learn==0.21.0
scipy==1.2.1
six==1.12.0
sklearn==0.0
